cleanup comments

+++ src/lib/particle_pack.h
@@ -2,9 +2,8 @@
 #define NB_PARTICLE_PACK_H
 
 #include <nbody.h>
-#include <stddef.h>
 
-/* Some number of particles pack together for vectorization. */
+/* Some number of particles packed together for vectorization. */
 typedef struct ParticlePack ParticlePack;
 
 /* Allocate ParticlePack that can fit COUNT particles. */
@@ -14,6 +13,6 @@ void AllocPackArray(uint32_t count, ParticlePack **arr, uint32_t *len);
 void PackParticles(uint32_t count, const Particle *ps, ParticlePack *packs);
 
 /* Update P with PACKS. */
-void PackedUpdate(Particle *p, float dt, size_t packs_len, ParticlePack *packs);
+void PackedUpdate(Particle *p, float dt, uint32_t packs_len, ParticlePack *packs);
 
 #endif //NB_PARTICLE_PACK_H
